# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorOS API
  version: 0.10.0
  description: REST contract for the CreatorOS modular monolith.
servers:
  - url: http://localhost:8080
    description: Local API
security: []
...
paths:
...
  /api/v1/orders/{order_id}/submissions:
    get:
      operationId: getOrderWorkflow
      summary: List submissions and workflow summary for an order
      tags: [Workflow]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Workflow summary and submission history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "404":
          $ref: "#/components/responses/OrderNotFound"
    post:
      operationId: submitContentDeliverables
      summary: Submit content deliverables for an order
      tags: [Workflow]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContentRequest"
      responses:
        "201":
          description: Content submission created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/InvalidOrderTransition"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/orders/{order_id}/submissions/upload:
    post:
      operationId: uploadDeliverableFile
      summary: Upload a deliverable file for an order submission
      tags: [Workflow]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Deliverable file uploaded to server storage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedFileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/orders/{order_id}/submissions/{submission_id}/revision:
    post:
      operationId: requestSubmissionRevision
      summary: Request a revision for a submission
      tags: [Workflow]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/SubmissionID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionRequest"
      responses:
        "200":
          description: Revision requested successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/SubmissionNotFound"
        "409":
          $ref: "#/components/responses/RevisionLimitExceeded"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/orders/{order_id}/submissions/{submission_id}/approve:
    post:
      operationId: approveSubmission
      summary: Approve submission and complete order
      tags: [Workflow]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/SubmissionID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveContentRequest"
      responses:
        "200":
          description: Submission approved and order completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/SubmissionNotFound"
        "409":
          $ref: "#/components/responses/InvalidOrderTransition"
  /api/v1/submissions/files/{file_id}:
    get:
      operationId: downloadDeliverableFile
      summary: Download or stream deliverable file
      tags: [Workflow]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: Deliverable binary file stream.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "404":
          $ref: "#/components/responses/FileNotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: creatoros_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must match the creatoros_csrf cookie bound to the current session.
      schema:
        type: string
...
    OrderID:
      name: order_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SubmissionID:
      name: submission_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    FileID:
      name: file_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The JSON body is malformed, too large, contains unknown fields, or is not a single object.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: One or more request fields are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthenticated:
      description: A valid active session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidCSRF:
      description: The CSRF cookie and header are missing, different, or not bound to the session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenAction:
      description: The authenticated role or permission cannot perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    OrderNotFound:
      description: The order was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    InvalidOrderTransition:
      description: The order cannot transition from its current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    SubmissionNotFound:
      description: The submission was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RevisionLimitExceeded:
      description: The order has reached its allowed revision limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    FileNotFound:
      description: The requested deliverable file was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PayloadTooLarge:
      description: The uploaded file exceeds the 200MB size limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnsupportedMediaType:
      description: The uploaded file MIME type is not allowed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    SubmissionStatus:
      type: string
      enum:
        - submitted
        - revision_requested
        - approved
    SubmissionFile:
      type: object
      required:
        - id
        - submission_id
        - file_name
        - file_size_bytes
        - mime_type
        - file_path
        - download_url
        - created_at
      properties:
        id: { type: string, format: uuid }
        submission_id: { type: string, format: uuid }
        file_name: { type: string }
        file_size_bytes: { type: integer, format: int64 }
        mime_type: { type: string }
        file_path: { type: string }
        download_url: { type: string }
        created_at: { type: string, format: date-time }
    SubmissionRevision:
      type: object
      required:
        - id
        - submission_id
        - revision_number
        - feedback
        - requested_by
        - created_at
      properties:
        id: { type: string, format: uuid }
        submission_id: { type: string, format: uuid }
        revision_number: { type: integer, minimum: 1 }
        feedback: { type: string }
        requested_by: { type: string, format: uuid }
        requested_by_name: { type: string }
        created_at: { type: string, format: date-time }
    Submission:
      type: object
      required:
        - id
        - order_id
        - creator_user_id
        - version
        - status
        - title
        - notes
        - files
        - revisions
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        creator_user_id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        status: { $ref: "#/components/schemas/SubmissionStatus" }
        title: { type: string }
        notes: { type: string }
        files:
          type: array
          items: { $ref: "#/components/schemas/SubmissionFile" }
        revisions:
          type: array
          items: { $ref: "#/components/schemas/SubmissionRevision" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkflowSummary:
      type: object
      required:
        - order_id
        - order_status
        - revision_limit
        - revision_count
        - submissions
      properties:
        order_id: { type: string, format: uuid }
        order_status: { type: string }
        revision_limit: { type: integer, minimum: 0 }
        revision_count: { type: integer, minimum: 0 }
        submissions:
          type: array
          items: { $ref: "#/components/schemas/Submission" }
    SubmitFileInput:
      type: object
      required:
        - file_path
        - file_name
        - mime_type
        - file_size_bytes
      properties:
        file_path: { type: string }
        file_name: { type: string }
        mime_type: { type: string }
        file_size_bytes: { type: integer, format: int64 }
    SubmitContentRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - files
      properties:
        title: { type: string, minLength: 3, maxLength: 160 }
        notes: { type: string, maxLength: 2000 }
        files:
          type: array
          minItems: 1
          maxItems: 10
          items: { $ref: "#/components/schemas/SubmitFileInput" }
    RevisionRequest:
      type: object
      additionalProperties: false
      required:
        - feedback
      properties:
        feedback: { type: string, minLength: 10, maxLength: 2000 }
    ApproveContentRequest:
      type: object
      additionalProperties: false
      properties:
        note: { type: string, maxLength: 1000 }
    UploadedFile:
      type: object
      required:
        - file_path
        - file_name
        - mime_type
        - file_size_bytes
      properties:
        file_path: { type: string }
        file_name: { type: string }
        mime_type: { type: string }
        file_size_bytes: { type: integer, format: int64 }
    UploadedFileResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/UploadedFile" }
    SubmissionResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Submission" }
    WorkflowSummaryResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/WorkflowSummary" }
...